+++ src/views/remotesNode.ts
@@ -12,9 +12,14 @@ export class RemotesNode extends ExplorerNode {
         constructor(
             uri: GitUri,
             private readonly repo: Repository,
-            private readonly explorer: GitExplorer
+            private readonly explorer: GitExplorer,
+            private readonly active: boolean = false
         ) {
             super(uri);
+        }
+
+        get id(): string {
+            return `gitlens:repository(${this.repo.path})${this.active ? ':active' : ''}:remotes`;
         }
 
         async getChildren(): Promise<ExplorerNode[]> {