+++ src/views/stashesNode.ts
@@ -11,9 +11,14 @@ export class StashesNode extends ExplorerNode {
     constructor(
         uri: GitUri,
         private readonly repo: Repository,
-        private readonly explorer: Explorer
+        private readonly explorer: Explorer,
+        private readonly active: boolean = false
     ) {
         super(uri);
+    }
+
+    get id(): string {
+        return `gitlens:repository(${this.repo.path})${this.active ? ':active' : ''}:stashes`;
     }
 
     async getChildren(): Promise<ExplorerNode[]> {