+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Fixed
 - Fixes [#314](https://github.com/eamodio/vscode-gitlens/issues/314) - Toggle line annotation doesn't work properly
 - Fixes [#310](https://github.com/eamodio/vscode-gitlens/issues/310) - "via Terminal" commands need quoting around work directory
+- Fixes issues with the active repository in the *GitLens* explorer failed to update properly
 - Fixes issues with *Open File*, *Open Revision*, *Show File History* commands and images and other binary files
 
 ## [8.1.1] - 2018-03-12